web/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.0.2
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
gunicorn==21.2.0
packaging==23.2
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2024.1
sqlparse==0.4.2
typing_extensions==4.9.0
